remake example to match real world app
